Replace int by double|

+++ common/Timer.hpp
@@ -50,7 +50,7 @@ public:
    * Returns the elapsed time in milliseconds since the reference time.
    * Optionally resets the reference time.
    */
-  int64_t msecs(reset_t reset = keepRunning);
+  double msecs(reset_t reset = keepRunning);
 
   /**
    * Returns the elapsed time in seconds (with microsecond precision).
